update code 202308-04

+++ backend/app/Helpers/GoogleDriver.php
@@ -35,6 +35,8 @@ class GoogleDriver
 
     public static function delete($image)
     {
-        Storage::disk('google')->delete($image);
+        if($image) {
+            Storage::disk('google')->delete($image);
+        }
     }
 }
